rag bot/requirements.txt
python-dotenv
llama-index
llama-index-llms-groq
llama-index-core
llama-index-readers-file
llama-index-tools-wolfram-alpha
llama-index-embeddings-huggingface
llama-index-embeddings-voyageai
crewai_tools==0.4.26
voyageai